Improvements

+++ client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@ import { Container } from '@angular/compiler/src/i18n/i18n_ast';
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -18,7 +19,7 @@ export class MessagesComponent implements OnInit {
   loading = false;
 
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService:ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -34,9 +35,14 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+    this.confirmService.confirm('Confirm delete message','This cannot be undone').subscribe(result=>{
+      if(result){
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+        })
+      }
     })
+   
   }
 
   pageChanged(event: any) {
